+++ lib/src/auth/domain/repository/auth_repository.dart
@@ -21,4 +21,6 @@ abstract interface class AuthRepository {
   Future<Either<Failure, void>> signOut();
 
   Future<Either<Failure, void>> deleteAccount();
+
+  Future<Either<Failure, void>> sendPasswordResetEmail(String email);
 }